Another analysis
Here's an analysis I've been working on recently. It's still a work in progress, but I'm trying to show that topic space can help us predict whether or not withheld questions will be answered correctly. (This fits our narrative that topic space is a useful way of understanding a person's knowledge). I know we've done some similar analyses, but this approach is a bit different in a few ways. One such way is that, as a point of comparison, I demonstrate how accurately we could predict the correctness of held-out questions if we didn't use topic space -- that is, if all we knew was the percentage of the questions (excluding the held-out one) that were answered correctly.